feat: add citation inline mode
