+++ libs/kotaemon/kotaemon/indices/qa/__init__.py
@@ -1,7 +1,5 @@
 from .citation import CitationPipeline
-from .text_based import CitationQAPipeline
 
 __all__ = [
     "CitationPipeline",
-    "CitationQAPipeline",
 ]
